Remove fundamentally broken symlink tests
These tests depend on being able to lchown a symbolic link to /bin/sh,
and rely on the assumption that /bin/sh is not owned by the current
user.

The first assumption will typically fail on Linux whenever the second
assumption is true!

Fixing the first, by linking to a local file instead, means that the
second is violated.

There is no way to save these tests, so they are being removed.